new think

+++ 07_02_2024/Task_1.py
@@ -1,5 +1,5 @@
 # Write a Python program to calculate the area of a circle given
 # its radius using the formula area=π×r^2 ( Take pie as 3.14)
-r = int(input("Please enter radius of circle"))
+r = float(input("Please enter radius of circle"))
 area = float(3.14 * (r ** 2))
 print("Area of circle you have enter is = ", area)
